[#32096] PG19: Fixes for yb.orig.create_table
Summary:
- src/postgres/src/backend/catalog/yb_catalog/yb_catalog_version.c:
  - YbGetSQLIncrementCatalogVersionFunctionOidHelper:
    - PG commit 83a56419457ec0eff2eddfed8eb3aba86bede9cc added an `int *fgc_flags` outparam
      to FuncnameGetCandidates and writes *fgc_flags unconditionally
      (see namespace.c). The PG19 initial merge commit 4c01ae04c6d8e5487679d2079e44a067a66febcd
      adapted the YB call by passing NULL, leading to a NULL dereference.
      bug cannot occur (PG behavior change). All in-tree PG callers pass a real &local.
    - Passed a throwaway `int fgc_flags;` local instead of NULL.

- src/postgres/src/backend/commands/yb_cmds.c:
  - YbBackfillIndexResultDesc:
    - PG commit c456e39113809376f6604e720910ccd24e18e034
      added TupleDesc.firstNonCachedOffsetAttr (initialized to -1) + compact_attrs, populated by the
      new TupleDescFinalize(), and added Assert(tupleDesc->firstNonCachedOffsetAttr >= 0) to the
      deform paths (execTuples.c slot_deform_heap_tuple etc.).
    - Added a TupleDescFinalize(tupdesc) call.

- src/postgres/src/backend/utils/cache/funccache.c, src/include/utils/funccache.h:
  - cached_function_compile / CachedFunction struct:
    - PG commit 0dca5d68d7bebf2c1036fd84875533afef6df992 centralized SQL/PL function compilation into the new
      funccache.c. cached_function_compile caches the pg_proc tuple's fn_xmin + fn_tid (heap
      ctid). YB does not use heap ctids, so the cache's ctid-based validation hits assertion failures.
    - Mirrored the fmgr.c pattern -- added uint64 yb_catalog_version to
      CachedFunction (funccache.h); in the validity check, use
      yb_catalog_version == YBGetActiveCatalogCacheVersion() under IsYugaByteEnabled().
      Added #include "pg_yb_utils.h".

- src/postgres/src/backend/utils/misc/guc_parameters.dat:
  - yb_explain_hide_non_deterministic_fields, yb_enable_create_with_table_oid:
    - PG commit 0a20ff54f5e66158930d5328f89f087d4e9ab400 moved GUC entries into this .dat file
      (consumed by gen_guc_tables.pl). The YB ConfigureNames* arrays in guc.c are #if 0'd for
      now, and need to be migrated (see guc.c).
    - Migrated yb_explain_hide_non_deterministic_fields and yb_enable_create_with_table_oid
      (alphabetical order as per PG). BasePgRegressTest sets yb_explain_hide_non_deterministic_fields
      to true for pg regress tests, and yb_enable_create_with_table_oid is used in yb.orig.create_table.

- src/postgres/src/include/access/itup.h:
  - IndexInfoFindDataOffset:
    - PG commit 14a8bd98274aaf51912b29cbb869696fc825da00 turned IndexInfoFindDataOffset from a macro
      into a static inline function taking `unsigned short t_info` (correct for upstream's uint16
      IndexTupleData.t_info). YB had widened t_info from unsigned short to uint32 to allow
      large index tuples (YB commit 2860a1b634b786d418d3ae9744e74b03b139a903).
      On YB master IndexInfoFindDataOffset is still a MACRO, so it sees the full
      uint32; the PG19 initial merge brought in PG's inline-function signature, so passing
      the uint32 t_info to the `unsigned short` parameter truncates to 16 bits.
    - Widened the IndexInfoFindDataOffset parameter from `unsigned short` to
      `uint32` to match YB's widened t_info.

Additionally, this revision fixes a lint issue in yb_cmds.c (no YB includes in yb_ file) that was introduced by the
initial merge commit.

Test Plan:
Tested locally:

- Reduced `yb_table_schedule` to just `yb.orig.create_table` and `yb.orig.drop_table`.
- Ran ./yb_build.sh debug --java-test 'org.yb.pgsql.TestPgRegressTable' with the trimmed schedule.

Reviewers: patnaik.balivada

Reviewed By: patnaik.balivada

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D54281

+++ src/postgres/src/include/access/itup.h
@@ -122,9 +122,12 @@ extern IndexTuple index_truncate_tuple(TupleDesc sourceDescriptor,
 /*
  * Takes an infomask as argument (primarily because this needs to be usable
  * at index_form_tuple time so enough space is allocated).
+ *
+ * YB: t_info is uint32 here (not PG's unsigned short) to match the
+ * widened IndexTupleData.t_info.  Widened because YB does not use TOAST.
  */
 static inline Size
-IndexInfoFindDataOffset(unsigned short t_info)
+IndexInfoFindDataOffset(uint32 t_info) /* YB */
 {
 	if (!(t_info & INDEX_NULL_MASK))
 		return MAXALIGN(sizeof(IndexTupleData));
